requirements.txt
python-dotenv
tqdm
numpy
pandas
opencv-python
matplotlib